fixed exception in cases on which the website could not been downloaded

+++ src/main/java/org/foppiano/education/crawler/link/LinkParser.java
@@ -54,8 +54,9 @@ public class LinkParser {
                     System.err.println("Something wrong happened. Skipping.");
                 }
             }
-            page.getLocalFile().deleteOnExit();
-
+            if(page.getLocalFile() != null) {
+                page.getLocalFile().deleteOnExit();
+            }
 
         }
